Create H5Dread Test
This should be a simple test of H5D.read(). It creates a dataspace and a dataset, writes a string and tries to read the string back. I'm new to HDF5, so maybe I'm doing something wrong but this test is not working for me. I always receive -1 from the H5D.read().